# omicsdm-server/server/swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "paths": {
...
    "/api/projects/": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "type": "string",
          "required": true
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Update a project",
        "operationId": "post_project_update",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/admin/update": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "type": "string",
          "required": true
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Update a project",
        "operationId": "post_project_update",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/admin/view": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "type": "string",
          "required": true
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Get all the project data",
        "operationId": "post_project_data_admin_view",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/adminviewcols": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "type": "string",
          "required": true
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server error"
          }
        },
        "summary": "Get project submission fields",
        "description": "Returns...",
        "operationId": "post_project_view_cols",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/all": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "type": "string",
          "required": true
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Get all the projects data by keycloak group",
        "operationId": "post_project_list",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/create": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "type": "string",
          "required": true
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server error"
          }
        },
        "summary": "Create a new project",
        "description": "Returns...",
        "operationId": "post_project_create",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/submissioncols": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server error"
          }
        },
        "summary": "Get project submission fields",
        "description": "Returns...",
        "operationId": "get_project_submission_cols",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/validate": {
      "parameters": [
        {
          "name": "Authorization",
          "in": "header",
          "type": "string",
          "required": true
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server error"
          }
        },
        "description": "Returns...",
        "operationId": "post_project_creation_validation",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "version": "1.0"
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
...
  "host": "localhost"
}